Check that the path exists before remove
Useful when we are trying wxobj which are destroyed earlier.
wixobj are used when using Fragment instead of Module.

+++ cerbero/packages/wix_packager.py
@@ -170,7 +170,8 @@ class MSIPackager(PackagerBase):
         if not keep_temp:
             for msms in self.merge_modules.values():
                 for p in msms:
-                    os.remove(p)
+                    if os.path.exists(p):
+                      os.remove(p)
 
         return paths
